edge_camera/pyproject.toml
[build-system]
requires = ["setuptools>=66", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "edge_camera"
version = "1.0.0"
description = "vio-edge: camera service — standalone HTTP capture service"
requires-python = ">=3.9"
dependencies = [
    "fastapi[all]==0.115.6",
    "Pillow==9.4.0",
]

[project.optional-dependencies]
raspberry = ["picamera2==0.3.25"]
basler = ["pypylon==4.0.0"]
dev = [
    "autoflake==2.3.1",
    "behave==1.3.3",
    "black==24.10.0",
    "flake8==7.1.1",
    "isort==5.13.2",
    "pytest==8.3.4",
    "pytest-asyncio==0.25.3",
    "pytest-cov==5.0.0",
    "httpx==0.28.1",
]

[project.scripts]
edge_camera = "edge_camera.__main__:run"

[tool.setuptools.packages.find]
where = ["src/"]

[tool.black]
line-length = 120
target-version = ["py310", "py311", "py312", "py313"]

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
testpaths = "tests"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
norecursedirs = [".git", "venv/", ".pytest_cache/"]

[tool.behave]
junit = true
junit_directory = "reports/behave/"
paths = ["tests/functional_tests"]
